# openAPI.yaml
openapi: 3.0.3
info:
  title: Vehicle Management API
  version: 2.0.0
  description: |
    Secure and scalable API for managing users, vehicles, assignments, and maintenance operations.

    ## Features
    - RESTful endpoints with consistent plural naming
    - Comprehensive error handling
    - Input validation and sanitization
    - Pagination support
    - Security headers and rate limiting

    ## Authentication
    Currently using basic security measures. OAuth2/JWT implementation pending.
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.vehiculos.example.com
    description: Production server
paths:
...
  /reservations:
    get:
      summary: Get all reservations with optional filters
      description: |
        Get all reservations with optional filtering by user or vehicle.

        **Filter Parameters:**
        - `userId`: Filter reservations by specific user ID
        - `vehicleId`: Filter reservations by specific vehicle ID
        - Pagination is supported with `page` and `limit` parameters

        **Examples:**
        - `/reservations?userId=550e8400-e29b-41d4-a716-446655440000`
        - `/reservations?vehicleId=650e8400-e29b-41d4-a716-446655440001`
        - `/reservations?page=1&limit=5`
      tags: [Reservations]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number for pagination
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
          description: Number of items per page
        - name: userId
          in: query
          schema:
            type: string
            format: uuid
          description: "Filter reservations by user ID"
        - name: vehicleId
          in: query
          schema:
            type: string
            format: uuid
          description: "Filter reservations by vehicle ID"
      responses:
        "200":
          description: List of reservations retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReservationListResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      summary: Create a new reservation
      tags: [Reservations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReservationInput"
      responses:
        "201":
          description: Reservation created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReservationResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ASSIGNMENTS
...
components:
  schemas:
...
    ProblemDetails:
      type: object
      description: RFC 7807 Problem Details for HTTP APIs
      properties:
        type:
          type: string
          format: uri
          description: A URI reference that identifies the problem type
          example: "https://example.com/problems/validation-error"
        title:
          type: string
          description: A short, human-readable summary of the problem type
          example: "Validation Failed"
        status:
          type: integer
          description: The HTTP status code
          example: 400
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence
          example: "The request contains invalid data"
        instance:
          type: string
          format: uri
          description: A URI reference that identifies the specific occurrence
          example: "/users/123"
      required:
        - type
        - title
        - status
    ValidationProblemDetails:
      allOf:
        - $ref: "#/components/schemas/ProblemDetails"
        - type: object
          properties:
            errors:
              type: array
              description: List of validation errors
              items:
                type: object
                properties:
                  field:
                    type: string
                    description: The field that failed validation
                    example: "email"
                  message:
                    type: string
                    description: The validation error message
                    example: "Invalid email format"
                  code:
                    type: string
                    description: The validation error code
                    example: "invalid_string"
...
    ApiResponse:
      type: object
      properties:
        status:
          type: string
          enum: [success, error]
        message:
          type: string
        data:
          type: object
        pagination:
          $ref: "#/components/schemas/Pagination"
    Pagination:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
        limit:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
        pages:
          type: integer
          minimum: 0
...
    Reservation:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "750e8400-e29b-41d4-a716-446655440002"
        vehicleId:
          type: string
          format: uuid
          example: "650e8400-e29b-41d4-a716-446655440001"
        userId:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        startDate:
          type: string
          format: date-time
          example: "2025-01-15T09:00:00Z"
        endDate:
          type: string
          format: date-time
          example: "2025-01-15T17:00:00Z"
      required:
        - id
        - vehicleId
        - userId
        - startDate
        - endDate
    ReservationInput:
      type: object
      properties:
        vehicleId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
      required:
        - vehicleId
        - userId
        - startDate
        - endDate
    ReservationResponse:
      allOf:
        - $ref: "#/components/schemas/ApiResponse"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/Reservation"
    ReservationListResponse:
      allOf:
        - $ref: "#/components/schemas/ApiResponse"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/Reservation"

    # ASSIGNMENT SCHEMAS
...
  responses:
...
    BadRequest:
      description: Bad Request - Invalid input data
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ValidationProblemDetails"
...
    InternalServerError:
      description: Internal server error
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
# Global security (to be implemented)
# security:
#   - BearerAuth: []